+++ docs/changelog.md
@@ -2,6 +2,7 @@
 
 ## 0.15.6
 * Fix animation complete race condition when multiple equations are automatically automated between forms in a slide navigator
+* Add long form division symbol for equations to be used with box
 
 ## 0.15.5
 * Fix rearrange crash when using gl text rendering of equations without any forms